Fix raspbian build (armhf) Closes #190

+++ src/jsonrpccpp/client/connectors/filedescriptorclient.cpp
@@ -18,7 +18,7 @@
 #include <algorithm>
 
 #define BUFFER_SIZE 64
-#define MAX_WRITE_SIZE (long unsigned int) 1024
+#define MAX_WRITE_SIZE 1024
 #ifndef DELIMITER_CHAR
 #define DELIMITER_CHAR char(0x0A)
 #endif //DELIMITER_CHAR
@@ -43,7 +43,7 @@ void FileDescriptorClient::SendRPCMessage(const std::string& message,
   string toSend = message;
   do
   {
-    ssize_t byteWritten = write(outputfd, toSend.c_str(), min(toSend.size(), MAX_WRITE_SIZE));
+    ssize_t byteWritten = write(outputfd, toSend.c_str(), min(toSend.size(), (size_t)MAX_WRITE_SIZE));
     if (byteWritten < 1)
       throw JsonRpcException(Errors::ERROR_CLIENT_CONNECTOR,
         "Unknown error occured while writing to the output file descriptor");
@@ -76,4 +76,3 @@ bool FileDescriptorClient::IsReadable(int fd)
   return ((o_accmode & O_ACCMODE) == O_RDONLY ||
     (o_accmode & O_ACCMODE) == O_RDWR);
 }
-
